Shownumber of events for the different time periods even if not shown
